requirements.txt
absl-py==2.1.0
accelerate==0.27.2
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
attrs==23.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
diffusers @ git+https://github.com/huggingface/diffusers@5266ab7935dd9e9aec596cdc2464badf1eacd99a
docker-pycreds==0.4.0
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.21.0
gradio_client==0.12.0
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
importlib_metadata==7.0.2
importlib_resources==6.1.3
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
orjson==3.9.15
packaging==24.0
pandas==2.2.1
peft==0.10.0
pillow==10.2.0
preshed==3.0.9
protobuf==4.25.3
psutil==5.9.8
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.2
safetensors==0.4.2
segment-anything==1.0
semantic-version==2.10.0
sentry-sdk==1.44.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.4
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.4
srsly==2.4.8
starlette==0.36.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thinc==8.2.3
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.36.2
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
wandb==0.16.5
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
websockets==11.0.3
Werkzeug==3.0.1
zipp==3.17.0
